Move fqs client to new services/fqs directory
